Remove old polymesh-common-utilities module.
