+++ pallets/asset/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 
 [dependencies]
 # Common
-polymesh-common-utilities = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 polymesh-primitives-derive = { workspace = true, default-features = false }
 
@@ -59,7 +58,6 @@ std = [
     "pallet-external-agents/std",
     "pallet-statistics/std",
     "polymesh-primitives/std",
-    "polymesh-common-utilities/std",
     "polymesh-runtime-common/std",
     "rustc-hex/std",
     "serde/std",
