+++ pallets/compliance-manager/src/lib.rs
@@ -87,12 +87,12 @@ use sp_std::{convert::From, prelude::*};
 
 use pallet_base::ensure_length_ok;
 use pallet_external_agents::{Config as EAConfig, GroupOfAgent};
-use polymesh_common_utilities::protocol_fee::{ChargeProtocolFee, ProtocolOp};
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::compliance_manager::{
     AssetCompliance, ComplianceReport, ComplianceRequirement, ConditionReport, RequirementReport,
 };
 use polymesh_primitives::condition::{conditions_total_counts, Condition};
+use polymesh_primitives::protocol_fee::{ChargeProtocolFee, ProtocolOp};
 use polymesh_primitives::traits::{AssetFnConfig, ComplianceFnConfig};
 use polymesh_primitives::{
     proposition, storage_migration_ver, Claim, ConditionType, Context, IdentityId, TargetIdentity,
