+++ pallets/corporate-actions/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 
 [dependencies]
 # Common
-polymesh-common-utilities = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 polymesh-primitives-derive = { workspace = true, default-features = false }
 
@@ -53,7 +52,6 @@ std = [
     "pallet-portfolio/std",
     "pallet-timestamp/std",
     "polymesh-primitives/std",
-    "polymesh-common-utilities/std",
     "serde/std",
     "serde_derive",
     "sp-api/std",
