+++ pallets/corporate-actions/src/lib.rs
@@ -106,8 +106,8 @@ use pallet_asset::checkpoint::{SchedulePoints, Timestamps};
 use pallet_asset::{checkpoint, BalanceOf};
 use pallet_base::try_next_post;
 use pallet_identity::{Config as IdentityConfig, PermissionedCallOriginData};
-use polymesh_common_utilities::checkpoint::ScheduleId;
 use polymesh_primitives::asset::{AssetId, CheckpointId};
+use polymesh_primitives::checkpoint::ScheduleId;
 use polymesh_primitives::{impl_checked_inc, storage_migration_ver, Balance, DocumentId};
 use polymesh_primitives::{EventDid, IdentityId, Moment, PortfolioNumber, GC_DID};
 use polymesh_primitives_derive::VecU8StrongTyped;
