+++ pallets/identity/src/claims.rs
@@ -19,8 +19,8 @@ use sp_runtime::traits::SaturatedConversion;
 use sp_std::prelude::*;
 
 use pallet_base::{ensure_string_limited, try_next_pre};
-use polymesh_common_utilities::protocol_fee::ProtocolOp;
 use polymesh_primitives::identity_claim::CustomClaimTypeId;
+use polymesh_primitives::protocol_fee::ProtocolOp;
 use polymesh_primitives::traits::group::GroupTrait;
 use polymesh_primitives::{
     CddId, Claim, ClaimType, IdentityClaim, IdentityId, Scope, SecondaryKey, SystematicIssuers,
