+++ pallets/pips/src/lib.rs
@@ -95,8 +95,8 @@ use sp_version::RuntimeVersion;
 use pallet_balances::impls::LockableCurrencyExt;
 use pallet_base::{ensure_opt_string_limited, try_next_post};
 use pallet_identity::{Config as IdentityConfig, PermissionedCallOriginData};
-use polymesh_common_utilities::protocol_fee::{ChargeProtocolFee, ProtocolOp};
 use polymesh_primitives::constants::PIP_MAX_REPORTING_SIZE;
+use polymesh_primitives::protocol_fee::{ChargeProtocolFee, ProtocolOp};
 use polymesh_primitives::storage_migration_ver;
 use polymesh_primitives::traits::group::GroupTrait;
 use polymesh_primitives::traits::GovernanceGroupTrait;
