+++ pallets/runtime/common/Cargo.toml
@@ -5,7 +5,6 @@ authors = ["PolymeshAssociation"]
 edition = "2021"
 
 [dependencies]
-polymesh-common-utilities = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 
 pallet-balances = { workspace = true, default-features = false  }
@@ -50,7 +49,6 @@ std = [
     "pallet-group-rpc-runtime-api/std",
     "pallet-identity/std",
     "pallet-multisig/std",
-    "polymesh-common-utilities/std",
     "polymesh-primitives/std",
     "sp-genesis-builder/std",
     "sp-io/std",
