+++ pallets/runtime/develop/Cargo.toml
@@ -10,7 +10,6 @@ build = "build.rs"
 [dependencies]
 # Common
 polymesh-runtime-common = { workspace = true, default-features = false }
-polymesh-common-utilities = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 polymesh-weights = { workspace = true, default-features = false }
