+++ pallets/runtime/mainnet/Cargo.toml
@@ -9,7 +9,6 @@ build = "build.rs"
 
 [dependencies]
 # Common
-polymesh-common-utilities = { workspace = true, default-features = false }
 polymesh-primitives = { workspace = true, default-features = false }
 polymesh-runtime-common = { workspace = true, default-features = false }
 polymesh-weights = { workspace = true, default-features = false }
@@ -172,7 +171,6 @@ std = [
     "pallet-utility/std",
     "polymesh-contracts/std",
     "polymesh-primitives/std",
-    "polymesh-common-utilities/std",
     "polymesh-runtime-common/std",
     "polymesh-weights/std",
     "serde/std",
