+++ pallets/runtime/tests/Cargo.toml
@@ -5,13 +5,6 @@ authors = ["PolymeshAssociation"]
 edition = "2021"
 
 [dependencies]
-# Common
-polymesh-runtime-common = { workspace = true, default-features = false }
-polymesh-runtime-develop = { workspace = true }
-polymesh-common-utilities = { workspace = true, default-features = false }
-polymesh-primitives = { workspace = true, default-features = false }
-polymesh-weights = { workspace = true, default-features = false }
-
 # Our pallets
 pallet-asset = { workspace = true, default-features = false }
 pallet-balances = { workspace = true, default-features = false }
@@ -40,6 +33,10 @@ polymesh-transaction-payment = { workspace = true, default-features = false }
 pallet-treasury = { workspace = true, default-features = false }
 pallet-utility = { workspace = true, default-features = false }
 polymesh-contracts = { workspace = true, default-features = false }
+polymesh-primitives = { workspace = true, default-features = false }
+polymesh-runtime-common = { workspace = true, default-features = false }
+polymesh-runtime-develop = { workspace = true }
+polymesh-weights = { workspace = true, default-features = false }
 polymesh-exec-macro = { path = "exec_macro" }
 
 # Others
@@ -173,7 +170,6 @@ std = [
     "pallet-treasury/std",
     "pallet-sudo/std",
     "polymesh-primitives/std",
-    "polymesh-common-utilities/std",
     "polymesh-runtime-common/std",
     "polymesh-contracts/std",
     "polymesh-weights/std",
