+++ pallets/runtime/tests/src/corporate_actions_test.rs
@@ -19,10 +19,10 @@ use pallet_corporate_actions::{CACheckpoint, CADetails, DefaultWithholdingTax};
 use pallet_corporate_actions::{CADocLink, CAId, CAKind, CorporateAction, DefaultTargetIdentities};
 use pallet_corporate_actions::{DidWithholdingTax, LocalCAId, MaxDetailsLength, RecordDate};
 use pallet_corporate_actions::{RecordDateSpec, TargetIdentities, TargetTreatment, Tax};
-use polymesh_common_utilities::checkpoint::{ScheduleCheckpoints, ScheduleId};
 use polymesh_primitives::agent::AgentGroup;
 use polymesh_primitives::asset::AssetId;
 use polymesh_primitives::asset::CheckpointId;
+use polymesh_primitives::checkpoint::{ScheduleCheckpoints, ScheduleId};
 use polymesh_primitives::constants::currency::ONE_UNIT;
 use polymesh_primitives::{AuthorizationData, Claim, ClaimType, Condition};
 use polymesh_primitives::{ConditionType, CountryCode, Document};
