+++ pallets/runtime/tests/src/protocol_fee.rs
@@ -4,7 +4,7 @@ use super::{
     ExtBuilder,
 };
 use frame_support::{assert_noop, assert_ok};
-use polymesh_common_utilities::protocol_fee::ProtocolOp;
+use polymesh_primitives::protocol_fee::ProtocolOp;
 use polymesh_primitives::traits::CddAndFeeDetails;
 use sp_keyring::Sr25519Keyring;
