+++ pallets/runtime/tests/src/storage.rs
@@ -31,8 +31,8 @@ use lazy_static::lazy_static;
 use pallet_identity::Context;
 use pallet_transaction_payment::RuntimeDispatchInfo;
 use pallet_utility;
-use polymesh_common_utilities::protocol_fee::ProtocolOp;
 use polymesh_primitives::constants::currency::{DOLLARS, POLY};
+use polymesh_primitives::protocol_fee::ProtocolOp;
 use polymesh_primitives::settlement::Leg;
 use polymesh_primitives::traits::{group::GroupTrait, CddAndFeeDetails};
 use polymesh_primitives::{AccountId, Authorization, AuthorizationData, BlockNumber};
