+++ primitives/src/identity.rs
@@ -16,8 +16,12 @@
 
 use serde::{Deserialize, Serialize};
 
-use codec::{Decode, Encode, MaxEncodedLen};
+use crate::{secondary_key::SecondaryKey, IdentityId};
+use codec::{Decode, DecodeWithMemTracking, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
+use sp_core::H512;
+
+pub type AuthorizationNonce = u64;
 
 #[cfg(feature = "running-ci")]
 /// Defines the constants for the identity pallet.
@@ -49,6 +53,49 @@ pub mod limits {
     pub const MAX_EXTRINSICS: usize = 80;
 }
 
+/// It represents an authorization that any account could sign to allow operations related with a
+/// target identity.
+///
+/// # Safety
+///
+/// Please note, that `nonce` has been added to avoid **replay attack** and it should be the current
+/// value of nonce of primary key of `target_id`. See `System::account_nonce`.
+/// In this way, the authorization is delimited to an specific transaction (usually the next one)
+/// of primary key of target identity.
+#[derive(Encode, Decode, Clone, PartialEq, Eq, Debug)]
+pub struct TargetIdAuthorization<Moment> {
+    /// Target identity which is authorized to make an operation.
+    pub target_id: IdentityId,
+    /// It HAS TO be `target_id` authorization nonce: See `Identity::offchain_authorization_nonce`
+    pub nonce: AuthorizationNonce,
+    pub expires_at: Moment,
+}
+
+/// Secondary key with authorization of that secondary key (off-chain operation) to be added
+/// to an identity.
+///
+/// `auth_signature` is the signature, generated by secondary key, of `TargetIdAuthorization`.
+#[derive(Decode, DecodeWithMemTracking, Encode, TypeInfo)]
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct SecondaryKeyWithAuth<AccountId> {
+    /// Secondary key to be added.
+    pub secondary_key: SecondaryKey<AccountId>,
+    /// Off-chain authorization signature.
+    pub auth_signature: H512,
+}
+
+/// Create a child identity using `key` as the primary key of the new child identity.
+///
+/// The `key` needs to sign (off-chain) an authorization.
+#[derive(Decode, DecodeWithMemTracking, Encode, TypeInfo)]
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub struct CreateChildIdentityWithAuth<AccountId> {
+    /// The key to be used as the primary key of a new child identity.
+    pub key: AccountId,
+    /// Off-chain authorization signature.
+    /// The signature is generated by `key` signing of `TargetIdAuthorization`.
+    pub auth_signature: H512,
+}
 /// Identity record.
 ///
 /// Used to check if an identity exists and lookup its primary key.
